second commit

+++ src/application/use-cases/employee_use-cases/delete-employee.use-case.ts
@@ -1,4 +1,5 @@
-import { Inject, Injectable } from '@nestjs/common';
+// src/application/use-cases/employee-use-cases/delete-employee.use-case.ts
+import { Injectable, Inject } from '@nestjs/common'; // @Inject импорт қилиш
 import { EmployeeRepository } from '../../../domain/repositories/employee.repository';
 
 @Injectable()
@@ -9,6 +10,11 @@ export class DeleteEmployeeUseCase {
   ) {}
 
   async execute(id: string): Promise<boolean> {
-    return this.employeeRepository.delete(id);
+    const employee = await this.employeeRepository.findById(id);
+    if (!employee) {
+      throw new Error('Employee not found');
+    }
+    await this.employeeRepository.delete(id);
+    return true;
   }
 }
